+++ app/jobs/[...slug]/page.tsx
@@ -1,2 +1,3 @@
-export { default, generateMetadata, revalidate } from '../_components/page'
+export const revalidate = 3600
 
+export { default, generateMetadata } from '../_components/page'